// src/app/models/marketing-management/banner/banner-item/banner-item.ts
export class BannerItem {
    _id: string;
    bannerItemName: string;
    bannerId: string;
    backgroundRGB: string;
    targetUrl: string;
    imageUrl: string;
    startDate: string;
    endDate: string;
    isDefault: boolean;
    status: number;
    statusName: string;
    updatedDate: string;
    createdDate: string;
    updatedUserId: string;
    createdUserId: string;
    priority: number
}